Update ImageController.java
Implemented Functionality to check If the owner of the image is trying to edit the image thereby enabling the user to edit the image.
But if the non-owner of the image is trying to edit the image,  print the message ‘Only the owner of the image can edit the image"

+++ ImageHoster/src/main/java/ImageHoster/controller/ImageController.java
@@ -45,11 +45,15 @@ public class ImageController {
     //Also now you need to add the tags of an image in the Model type object
     //Here a list of tags is added in the Model type object
     //this list is then sent to 'images/image.html' file and the tags are displayed
-    @RequestMapping("/images/{title}")
-    public String showImage(@PathVariable("title") String title, Model model) {
-        Image image = imageService.getImageByTitle(title);
+    @RequestMapping("/images/{imageId}/{title}")
+    public String showImage(@PathVariable(name = "imageId") Integer imageId, @PathVariable(name = "title") String title, Model model) {
+
+        //Changing method to getImageObject by passing imageId,title
+        Image image = imageService.getImageByTitle(imageId, title);
+
         model.addAttribute("image", image);
         model.addAttribute("tags", image.getTags());
+
         return "images/image";
     }
 
@@ -92,13 +96,26 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model,HttpSession session) {
         Image image = imageService.getImage(imageId);
+        // Code added to get the logged user from the session object and handle error message to be displayed to restrict editing of images
+        // by other users (not the user who uploaded the image).
 
-        String tags = convertTagsToString(image.getTags());
+        User user = (User) session.getAttribute("loggeduser");
+        Integer imageUserID = imageService.getImage(imageId).getUser().getId();      //getting User ID corresponding to the specified image.
+
         model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+        if (imageUserID != user.getId()){
+            model.addAttribute("tags", image.getTags());
+            String error = "Only the owner of the image can edit the image";
+            model.addAttribute("editError", error);
+            return "images/image";
+        }
+        else {
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -140,9 +157,28 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, Model model) {
+
+        // modified method to enable delete image functionality by the user who created it.
+        User user = (User) session.getAttribute("loggeduser"); // To get user object value for logged-in user
+        Image image = imageService.getImage(imageId); // Fetch image by id
+        User imageOwner = image.getUser(); // Using getUser function of Image class to get the current user
+
+        //Comparison to check if the logged-in user's id matches to the id of userId of the user who uploaded the image
+        // If true user can delete the image, else the error message.
+        if(user.getId() == imageOwner.getId()) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+
+        else {
+            String error = "Only the owner of the image can delete the image";
+            model.addAttribute("tags", image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("deleteError", error);
+            return "images/image";
+        }
+
     }
